+++ Source/Controllers/PositionController.cs
@@ -4,14 +4,12 @@
 
     public class PositionController : AddableBase, IUpdatable
     {
-        #region Fields
         private Rect _target;
         private readonly float _initialX, _initialY;
         private readonly GetPosition _getX, _getY;
         private float _totalElapsed;
-        #endregion
+        
 
-        #region Constructors
         public PositionController(Rect target, bool positionAbsolute, GetPosition getX, GetPosition getY)
         {
             _target = target;
@@ -23,19 +21,17 @@
             _getX = getX;
             _getY = getY;
         }
-        #endregion
+        
 
-        #region IUpdateable
         public bool Active { get; set; } = true;
 
-        #region Update
-        public virtual void Update(double elapsed)
+        public virtual void Update(float elapsed)
         {
             _totalElapsed += (float)elapsed;
             _target.X = _initialX + _getX(_totalElapsed);
             _target.Y = _initialY + _getY(_totalElapsed);
         }
-        #endregion
-        #endregion
+        
+        
     }
 }
