+++ Source/Core/IUpdatable.cs
@@ -3,6 +3,6 @@
     public interface IUpdatable
     {
         bool Active { get; set; }
-        void Update(double elapsedTime);
+        void Update(float elapsedTime);
     }
 }
