+++ Source/Graphics/Animation.cs
@@ -8,15 +8,13 @@
 
     public class Animation : Sprite, IUpdatable
     {
-        #region Fields
         private int[] _framesToPlay = new int[] { 0 };
         private int _playIndex = 0;
         private float _timeToNextFrame;
         private LoopType _looping;
         private Action _onComplete;
-        #endregion
+        
 
-        #region Constructors
         public Animation(string imgPath, Point size, int spriteSheetColumns, int spriteSheetRows, float animationStepTime = 0.1f)
             : this(imgPath, new Rect(size), spriteSheetColumns, spriteSheetRows, animationStepTime)
         {
@@ -37,13 +35,11 @@
         {
             AnimStepTime = animationStepTime;
         }
-        #endregion
+        
 
-        #region IUpdatable
         public bool Active { get; set; } = true;
 
-        #region Update
-        public virtual void Update(double elapsed)
+        public virtual void Update(float elapsed)
         {
             if (Playing && _framesToPlay.Length > 1)
             {
@@ -66,18 +62,16 @@
                 }
             }
         }
-        #endregion
-        #endregion
+        
+        
 
-        #region Properties
         public float AnimStepTime { get; set; }
 
         public bool Playing { get; set; }
 
         public int[] AllFrames => Enumerable.Range(0, TotalFrames).ToArray();
-        #endregion
+        
 
-        #region Methods
         public void PlayAllOnce(Action actionOnComplete)
         {
             Play(LoopType.OneShot, AllFrames);
@@ -86,7 +80,6 @@
         public void PlayAllOnce() => Play(LoopType.OneShot, AllFrames);
         public void PlayAllLooping() => Play(LoopType.Looping, AllFrames);
 
-        #region Play
         /// <summary>
         /// Fixes the Sprite with a single frame
         /// </summary>
@@ -94,17 +87,15 @@
         public void Play(int frame) => Play(LoopType.OneShot, frame);
 
         public void Play(LoopType loopType, params int[] frames) => PlayFrom(loopType, 0, AnimStepTime, frames);
-        #endregion
+        
 
-        #region PlayOnce
         public void PlayOnce(Action actionOnComplete, int startFrame, params int[] frames)
         {
             PlayFrom(LoopType.OneShot, 0, AnimStepTime, frames);
             _onComplete = actionOnComplete;
         }
-        #endregion
+        
 
-        #region PlayFrom
         public void PlayFrom(LoopType loopType, int fromIndex, float currentFrameRemainingTime, params int[] frames)
         {
             Playing = true;
@@ -114,9 +105,8 @@
             Frame = frames[fromIndex];
             _looping = loopType;
         }
-        #endregion
+        
 
-        #region OnAnimationComplete
         protected void OnAnimationComplete()
         {
             if (_onComplete != null)
@@ -124,8 +114,8 @@
 
             AnimationComplete?.Invoke(this, EventArgs.Empty);
         }
-        #endregion
-        #endregion
+        
+        
 
         public event EventHandler AnimationComplete;
     }
