+++ Source/Tasks/TaskGroup.cs
@@ -2,14 +2,12 @@
 {
     public class TaskGroup : ITask
     {
-        #region Fields
         private bool _isStarted = false;
         protected readonly List<Action> ActionsOnStart = new();
         protected readonly List<Func<bool>> CompletionConditions = new();
         protected readonly List<Action> ActionsOnComplete = new();
-        #endregion
+        
 
-        #region Constructors
         public TaskGroup(params ITask[] tasks)
         {
             if (tasks.Length == 0)
@@ -25,17 +23,12 @@
         {
             CurrentTask = firstTask;
         }
-        #endregion
-
-        #region Properties
+        
         public virtual bool Active { get; set; } = true;
         public ITask? CurrentTask { get; set; }
         public bool IsComplete => CurrentTask == null && CompletionConditions.All(c => c());
         public ITask? NextTask { get; set; } = null;
-        #endregion
-
-        #region Methods
-        #region Complete
+        
         public virtual void Complete()
         {
             ActionsOnComplete.ForEach(a => a());
@@ -44,16 +37,12 @@
 
             _isStarted = false;
         }
-        #endregion
-
-        #region Reset
+        
         public virtual void Reset()
         {
             _isStarted = false;
         }
-        #endregion
-
-        #region Start
+        
         public virtual void Start()
         {
             ActionsOnStart.ForEach(a => a());
@@ -62,10 +51,8 @@
 
             _isStarted = true; //include this both here and in update in case function is overridden
         }
-        #endregion
-
-        #region Update
-        public virtual void Update(double elapsed)
+        
+        public virtual void Update(float elapsed)
         {
             if (!_isStarted)
             {
@@ -84,8 +71,8 @@
                 }
             }
         }
-        #endregion
-        #endregion
+        
+        
 
         public event EventHandler TaskStarted, TaskCompleted;
     }
