+++ Source/Tweens/LinearMotion.cs
@@ -16,7 +16,7 @@
             _rangeY = endY - startY;
         }
 
-        public override void Update(double elapsed)
+        public override void Update(float elapsed)
         {
             base.Update(elapsed);
